+++ NEWS.md
@@ -36,6 +36,8 @@
 
 * The `metro` and `films` datasets have been updated. (#2026, #2027)
 
+* Add option to control the order of footnotes (@shannonhaughton, #2023)
+
 # gt 1.0.0
 
 ## Minor improvements and bug fixes